support locally pre-build skia
