+++ skia-bindings/build_support.rs
@@ -2,15 +2,15 @@
 
 pub mod android;
 pub mod binaries_config;
+pub mod bind_skia;
 pub mod cargo;
 pub mod clang;
 pub mod features;
 pub mod ios;
-pub mod llvm;
-pub mod build_skia;
-pub mod bind_skia;
-pub mod vs;
 pub mod xcode;
+
+#[cfg(feature = "build-from-source")]
+pub mod build_skia;
 
 #[cfg(feature = "binary-cache")]
 pub mod binary_cache;
